fix ctrl-r not working when app crashes
